+++ Zilon.Core/Zilon.Core/Schemes/SectorCellularAutomataMapFactoryOptionsSubScheme.cs
@@ -16,13 +16,4 @@ namespace Zilon.Core.Schemes
         [JsonProperty]
         public int ChanceToStartAlive { get; private set; }
     }
-
-    public sealed class SectorOpenMapFactoryOptionsSubScheme : SectorMapFactoryOptionsSubSchemeBase,
-        ISectorOpenMapFactoryOptionsSubScheme
-    {
-        public override SchemeSectorMapGenerator MapGenerator => SchemeSectorMapGenerator.Open;
-
-        [JsonProperty]
-        public int Size { get; private set; }
-    }
 }
